Move do_* methods into the related extrinsic module (commit_weights.py).
