+++ bittensor/core/subtensor.py
@@ -1937,65 +1937,6 @@ class Subtensor:
         return success, message
 
     # Community uses this method
-    @networking.ensure_connected
-    def _do_commit_weights(
-        self,
-        wallet: "Wallet",
-        netuid: int,
-        commit_hash: str,
-        wait_for_inclusion: bool = False,
-        wait_for_finalization: bool = False,
-    ) -> Tuple[bool, Optional[str]]:
-        """
-        Internal method to send a transaction to the Bittensor blockchain, committing the hash of a neuron's weights.
-        This method constructs and submits the transaction, handling retries and blockchain communication.
-
-        Args:
-            wallet (bittensor.wallet): The wallet associated with the neuron committing the weights.
-            netuid (int): The unique identifier of the subnet.
-            commit_hash (str): The hash of the neuron's weights to be committed.
-            wait_for_inclusion (bool, optional): Waits for the transaction to be included in a block.
-            wait_for_finalization (bool, optional): Waits for the transaction to be finalized on the blockchain.
-
-        Returns:
-            Tuple[bool, Optional[str]]: A tuple containing a success flag and an optional error message.
-
-        This method ensures that the weight commitment is securely recorded on the Bittensor blockchain, providing a
-        verifiable record of the neuron's weight distribution at a specific point in time.
-        """
-
-        @retry(delay=1, tries=3, backoff=2, max_delay=4, logger=logging)
-        def make_substrate_call_with_retry():
-            call = self.substrate.compose_call(
-                call_module="SubtensorModule",
-                call_function="commit_weights",
-                call_params={
-                    "netuid": netuid,
-                    "commit_hash": commit_hash,
-                },
-            )
-            extrinsic = self.substrate.create_signed_extrinsic(
-                call=call,
-                keypair=wallet.hotkey,
-            )
-            response = self.substrate.submit_extrinsic(
-                extrinsic,
-                wait_for_inclusion=wait_for_inclusion,
-                wait_for_finalization=wait_for_finalization,
-            )
-
-            if not wait_for_finalization and not wait_for_inclusion:
-                return True, None
-
-            response.process_events()
-            if response.is_success:
-                return True, None
-            else:
-                return False, format_error_message(response.error_message)
-
-        return make_substrate_call_with_retry()
-
-    # Community uses this method
     def reveal_weights(
         self,
         wallet: "Wallet",
@@ -2059,71 +2000,3 @@ class Subtensor:
                 retries += 1
 
         return success, message
-
-    # Community uses this method
-    @networking.ensure_connected
-    def _do_reveal_weights(
-        self,
-        wallet: "Wallet",
-        netuid: int,
-        uids: List[int],
-        values: List[int],
-        salt: List[int],
-        version_key: int,
-        wait_for_inclusion: bool = False,
-        wait_for_finalization: bool = False,
-    ) -> Tuple[bool, Optional[str]]:
-        """
-        Internal method to send a transaction to the Bittensor blockchain, revealing the weights for a specific subnet.
-        This method constructs and submits the transaction, handling retries and blockchain communication.
-
-        Args:
-            wallet (bittensor.wallet): The wallet associated with the neuron revealing the weights.
-            netuid (int): The unique identifier of the subnet.
-            uids (List[int]): List of neuron UIDs for which weights are being revealed.
-            values (List[int]): List of weight values corresponding to each UID.
-            salt (List[int]): List of salt values corresponding to the hash function.
-            version_key (int): Version key for compatibility with the network.
-            wait_for_inclusion (bool, optional): Waits for the transaction to be included in a block.
-            wait_for_finalization (bool, optional): Waits for the transaction to be finalized on the blockchain.
-
-        Returns:
-            Tuple[bool, Optional[str]]: A tuple containing a success flag and an optional error message.
-
-        This method ensures that the weight revelation is securely recorded on the Bittensor blockchain, providing transparency
-        and accountability for the neuron's weight distribution.
-        """
-
-        @retry(delay=1, tries=3, backoff=2, max_delay=4, logger=logging)
-        def make_substrate_call_with_retry():
-            call = self.substrate.compose_call(
-                call_module="SubtensorModule",
-                call_function="reveal_weights",
-                call_params={
-                    "netuid": netuid,
-                    "uids": uids,
-                    "values": values,
-                    "salt": salt,
-                    "version_key": version_key,
-                },
-            )
-            extrinsic = self.substrate.create_signed_extrinsic(
-                call=call,
-                keypair=wallet.hotkey,
-            )
-            response = self.substrate.submit_extrinsic(
-                extrinsic,
-                wait_for_inclusion=wait_for_inclusion,
-                wait_for_finalization=wait_for_finalization,
-            )
-
-            if not wait_for_finalization and not wait_for_inclusion:
-                return True, None
-
-            response.process_events()
-            if response.is_success:
-                return True, None
-            else:
-                return False, format_error_message(response.error_message)
-
-        return make_substrate_call_with_retry()
